requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pandas==2.1.1
openpyxl==3.1.2
python-dateutil==2.8.2
SQLAlchemy==2.0.21
Werkzeug==2.3.7
pytest==7.4.2
pytest-flask==1.2.0
pytest-cov==4.1.0
factory-boy==3.3.0
